% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot accidents for chosen state on a map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{state ID number}

\item{year}{value for year}
}
\value{
map plot
}
\description{
\code{fars_map_state} plots geographical distribution of accidents on a map
of respective state chosen by user using R's base graphics system. Function
stops with a message if state number is not found in a dataset or returns
NULL if there are no accidents in a given state.
}
\examples{

my_state_id <- 1
fars_map_state(my_state_id, 2015)

}
