% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Import data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{name of a dataset as character string}
}
\value{
data.frame of class: data.frame and tbl_df
}
\description{
\code{fars_read} imports data into R session as tbl_df.
}
\details{
This is a wrapper function, not very complicated, it only reads data from
disk. It stops if file is not found, returning a message. If it finds
the file on disk, it imports it, suppresssing messages generated by
\code{readr::read_csv} function. In the end, the function converts data to
\code{dplyr::tbl_df} even though conversion is already done by
\code{readr::read_csv} function.
}
\examples{
my_filename <- here::here("data/accident_2013.csv.bz2")
my_df <- fars_read(filename = my_filename)
my_df
}
