% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Import dataset for multiple years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{vector of years as integers}
}
\value{
a list where each element of a list is an object of class tbl_df
and data.frame
}
\description{
\code{fars_read_years} imports into R session a set of datasets at once,
returning a list of data.frames objects of class tbl_df, data.frame and tbl.
After importing, it adds column and fills it with a respective year number,
than selects columns Month and year.

If dataset names to import do not contain year values provided by user,
the function returns NULL with a warning.
}
\examples{

my_years <- 2013:2015
my_dataset_list <- fars_read_years(my_years)
my_dataset_list

}
