% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Count observations by year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{vector of years as integers}
}
\value{
object of class data.frame, tbl_df, tbl
}
\description{
\code{fars_summarize_years} imports into R session a set of datasets at once,
binds rows of elements of list into a data.frame object,
counts observations by year and month, spreads result by year and value pair,
returning a data.frame object of class tbl_df, data.frame and tbl.
}
\examples{

my_years <- 2013:2015
my_count_by_year_and_month_df <- fars_summarize_years(my_years)
my_count_by_year_and_month_df
}
