% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Construct file name for a given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{integer year value}
}
\value{
filename string of class character
}
\description{
\code{make_filename} constructs filename string of class character using
argument year that is provided by user.
}
\details{
This function constructs a string that is a filename of dataset that one
would like to import into R session. By providing value for year, it will
construct the name using \code{sprintf} function in an appropriate manner
for dataset to be imported.
}
\examples{
my_year <- 2013
my_filename <- make_filename(year = my_year)
my_filename
}
